+++ clickhouse_connect/driver/insert.py
@@ -1,12 +1,15 @@
-import datetime
-from typing import Iterable, Sequence, Optional, Any, Dict, NamedTuple, Generator
+from typing import Iterable, Sequence, Optional, Any, Dict, NamedTuple, Generator, Union
 
 from clickhouse_connect.datatypes.base import ClickHouseType
+from clickhouse_connect.datatypes.registry import get_from_name
 from clickhouse_connect.driver.common import SliceView
-from clickhouse_connect.driver.options import check_pandas, check_numpy
+from clickhouse_connect.driver.context import BaseQueryContext
+from clickhouse_connect.driver.options import np, pd
 from clickhouse_connect.driver.exceptions import ProgrammingError
 
 DEFAULT_BLOCK_SIZE = 16834
+ch_nano_dt_type = get_from_name('DateTime64(9)')
+np_nano_dt_type = np.dtype('datetime64[ns]') if np else None
 
 
 class InsertBlock(NamedTuple):
@@ -18,7 +21,7 @@ class InsertBlock(NamedTuple):
 
 
 # pylint: disable=too-many-instance-attributes
-class InsertContext:
+class InsertContext(BaseQueryContext):
     """
     Reusable Argument/parameter object for inserts.
     """
@@ -27,58 +30,62 @@ class InsertContext:
                  table: str,
                  column_names: Sequence[str],
                  column_types: Sequence[ClickHouseType],
-                 data: Sequence[Sequence[Any]] = None,
+                 data: Any = None,
                  column_oriented: bool = False,
                  settings: Optional[Dict[str, Any]] = None,
                  compression: Optional[str] = None,
+                 query_formats: Optional[Dict[str, str]] = None,
+                 column_formats: Optional[Dict[str, Union[str, Dict[str, str]]]] = None,
                  block_size: int = DEFAULT_BLOCK_SIZE):
+        super().__init__(settings, query_formats, column_formats)
         self.table = table
         self.column_names = column_names
         self.column_types = column_types
         self.column_oriented = column_oriented
-        self.settings = settings
         self.compression = compression
         self.block_size = block_size
         self.data = data
         self.insert_exception = None
 
     @property
+    def empty(self) -> bool:
+        return self._data is None
+
+    @property
     def data(self):
-        return self._data
+        return self._raw_data
 
     @data.setter
-    def data(self, data):
+    def data(self, data: Any):
+        self._raw_data = data
         self.current_block = 0
         self.current_row = 0
         self.row_count = 0
         self.column_count = 0
-        if data:
-            if self.column_oriented:
-                self._next_block_data = self._column_block_data
-                self._block_columns = [SliceView(column) for column in data]
-                self._block_rows = None
-                self.column_count = len(data)
-                self.row_count = len(data[0])
-            else:
-                self._next_block_data = self._row_block_data
-                self._block_rows = data
-                self._block_columns = None
-                self.row_count = len(data)
-                self.column_count = len(data[0])
+        self._data = None
+        if data is None or len(data) == 0:
+            return
+        if pd and isinstance(data, pd.DataFrame):
+            data = self._convert_pandas(data)
+            self.column_oriented = True
+        if np and isinstance(data, np.ndarray):
+            data = self._convert_numpy(data)
+        if self.column_oriented:
+            self._next_block_data = self._column_block_data
+            self._block_columns = [SliceView(column) for column in data]
+            self._block_rows = None
+            self.column_count = len(data)
+            self.row_count = len(data[0])
+        else:
+            self._next_block_data = self._row_block_data
+            self._block_rows = data
+            self._block_columns = None
+            self.row_count = len(data)
+            self.column_count = len(data[0])
         if self.row_count and self.column_count:
             if self.column_count != len(self.column_names):
                 raise ProgrammingError('Insert data column count does not match column names')
             self._data = data
-        else:
-            self._data = None
-
-    @property
-    def df(self):
-        raise NotImplementedError('DataFrame is read only')
-
-    @df.setter
-    def df(self, df):
-        self.data = from_pandas_df(df, self.column_types)
 
     def next_block(self) -> Generator[InsertBlock, None, None]:
         while True:
@@ -98,35 +105,49 @@ class InsertContext:
         column_slice = SliceView(self._block_rows[block_start: block_end])
         return tuple(zip(*column_slice))
 
-    def __enter__(self):
-        return self
+    def _convert_pandas(self, df):
+        data = []
+        for df_col_name, col_name, ch_type in zip(df.columns, self.column_names, self.column_types):
+            df_col = df[df_col_name]
+            d_type = str(df_col.dtype)
+            if ch_type.python_type == int:
+                if 'float' in d_type:
+                    df_col = df_col.round().astype(ch_type.base_type, copy=False)
+                else:
+                    df_col = df_col.astype(ch_type.base_type, copy=False)
+            elif 'datetime' in ch_type.np_type() and np.issubdtype(df_col.dtype, np.datetime64):
+                div = ch_type.nano_divisor
+                data.append([None if pd.isnull(x) else x.value // div for x in df_col])
+                self.column_formats[col_name] = 'int'
+                continue
+            if ch_type.nullable and ch_type.python_type != float:
+                df_col.replace({np.nan: None}, inplace=True)
+            data.append(df_col.tolist())
+        return data
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        pass
+    def _convert_numpy(self, np_array):
+        if np_array.dtype.names is None:
+            if 'date' in str(np_array.dtype):
+                for col_name, col_type in zip(self.column_names, self.column_types):
+                    if 'date' in col_type.np_type():
+                        self.column_formats[col_name] = 'int'
+                return np_array.astype('int').tolist()
+            for col_type in self.column_types:
+                if col_type.byte_size == 0 or col_type.byte_size > np_array.dtype.itemsize:
+                    return np_array.tolist()
+            return np_array
 
-
-def from_pandas_df(df, column_types: Sequence[ClickHouseType]):
-    """
-    Convert a pandas dataframe into Python native types for insert
-    :param df: Pandas DataFrame for insert
-    :param column_types: ClickHouse column types matching the data frame columns in order.  Used for transformation
-    :return: Tuple of the dataframe column names and
-    """
-    np = check_numpy()
-    pd = check_pandas()
-    data = []
-    for col_name, ch_type in zip(df.columns, column_types):
-        df_col = df[col_name]
-        d_type = str(df_col.dtype)
-        if ch_type.python_type == int:
-            if 'float' in d_type:
-                df_col = df_col.round().astype(ch_type.base_type, copy=False)
-            else:
-                df_col = df_col.astype(ch_type.base_type, copy=False)
-        elif ch_type.python_type in (datetime.datetime, datetime.date) and 'date' in d_type:
-            data.append([None if pd.isnull(x) else pd.Timestamp.to_pydatetime(x) for x in df_col])
-            continue
-        if ch_type.nullable and ch_type.python_type != float:
-            df_col.replace({np.nan: None}, inplace=True)
-        data.append(df_col.tolist())
-    return data
+        if set(self.column_names).issubset(set(np_array.dtype.names)):
+            data = [np_array[col_name] for col_name in self.column_names]
+        else:
+            # Column names don't match, so we have to assume they are in order
+            data = [np_array[col_name] for col_name in np_array.dtype.names]
+        for ix, (col_name, col_type) in enumerate(zip(self.column_names, self.column_types)):
+            d_type = data[ix].dtype
+            if 'date' in str(d_type) and 'date' in col_type.np_type():
+                self.column_formats[col_name] = 'int'
+                data[ix] = data[ix].astype(int).tolist()
+            elif col_type.byte_size == 0 or col_type.byte_size > d_type.itemsize:
+                data[ix] = data[ix].tolist()
+        self.column_oriented = True
+        return data