+++ tests/unit_tests/test_superset/conftest.py
@@ -1,16 +1,24 @@
 import os
 import pathlib
-import sys
 
 from typing import Iterator
 
 from unittest import mock
 from pytest_mock import MockFixture
 from pytest import fixture
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 
-if sys.version_info[1] not in (8, 9):
+try:
+    from sqlalchemy import create_engine
+    from sqlalchemy.orm import sessionmaker
+    from superset.app import SupersetApp
+    from superset.initialization import SupersetAppInitializer
+except ImportError:
+    create_engine = None
+    sessionmaker = None
+    SupersetApp = None
+    SupersetAppInitializer = None
+
+if SupersetApp is None:
     collect_ignore_glob = ['test*.py']
 
 
@@ -22,20 +30,17 @@ def mock_settings_env_vars() -> Iterator[None]:
 
 
 @fixture(scope='session')
-# pylint: disable=import-outside-toplevel
 def superset_app(session_mocker: MockFixture) -> Iterator[None]:
-    from superset.app import SupersetApp
-    from superset.initialization import SupersetAppInitializer
+    if SupersetApp:
+        session = sessionmaker(bind=create_engine('sqlite://'))()
+        session.remove = lambda: None
 
-    session = sessionmaker(bind=create_engine('sqlite://'))()
-    session.remove = lambda: None
+        session_mocker.patch('superset.security.SupersetSecurityManager.get_session', return_value=session)
+        session_mocker.patch('superset.db.session', session)
 
-    session_mocker.patch('superset.security.SupersetSecurityManager.get_session', return_value=session)
-    session_mocker.patch('superset.db.session', session)
+        app = SupersetApp(__name__)
+        app.config.from_object('superset.config')
 
-    app = SupersetApp(__name__)
-    app.config.from_object('superset.config')
-
-    SupersetAppInitializer(app).init_app()
-    with app.app_context():
-        yield
+        SupersetAppInitializer(app).init_app()
+        with app.app_context():
+            yield